refactor(blog): convert from in-browser Babel to Vite static build
Replace CDN React/Babel with proper ES modules + Vite bundler:
- IIFE + window globals → ES module imports/exports
- window.Blog.register() → export default { id, Component, meta }
- Assets moved to public/ for static serving
- Add GitHub Actions workflow for build + deploy to Pages

Build output: 196 KB JS + 35 KB CSS (67 KB + 7 KB gzipped).
