+++ blog/src/main.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './themes.css';
+import './posts/index.js';
+import App from './shell-ui';
+
+ReactDOM.createRoot(document.getElementById('root')).render(<App />);
